update examples and documentation
