+++ docs/stripe /index.md
@@ -1,4 +1,4 @@
-# Proxy
+# Stripe
 Proxy can be used when you wish to make server side requests from your player to one of the enabled integrations on your account.
 
 Proxy uses standard window.fetch under the hood. The second argument (request) should match the [RequestInit structure](https://github.com/microsoft/TypeScript/blob/main/lib/lib.webworker.d.ts#L436-L489). The only difference is by default the method will use `'Content-Type': 'application/json'`. This and any other options as part of the `RequestInit` can be overridden.
@@ -15,7 +15,7 @@ You can view a [working example here](./src).
 try {
 
   // First segment is the name, remaining segments will be the endpoints for the related platform
-  const url = '/square'
+  const url = '/stripe'
 
   // Default method is a GET and by default the player will use 'Content-Type': 'application/json'. All RequestInit options can be overridden using the second argument.
   const request: RequestInit = {
