+++ docs/stripe /src/index.html
@@ -2,19 +2,17 @@
 <html>
 
 <head>
-  <title>Evexi API Demo - Proxy</title>
+  <title>Evexi API Demo - Stripe</title>
   <style>
     body {
-      background-color: lightsalmon;
+      background-color: burlywood;
       transition: background-color ease-in-out 1s;
     }
   </style>
 </head>
 
 <body>
-  <h2>Evexi API Demo Proxy</h2>
-
-  <br />
+  <h2>Evexi API Demo Stripe</h2>
 
   <div id="logs"></div>
   <script type="module" src="index.ts"></script>
